Changed API to read from gcc map file & added erase
I realized that the bss wasn't being initialized. I then also realized
that the api status data wasn't being initialized either, so it could
be anything. So, I made the python stuff parse a map file instead,
looking for the addresses of specific sections in order to find the
pertinent pieces of code to run

I added the mass erase command...it hasn't been tested yet